Rename `server` to `api`
