+++ packages/api/src/index.ts
@@ -1,4 +1,4 @@
-import {REPO_DIR, SERVER_PORT} from '@internal/constants';
+import {API_PORT, REPO_DIR} from '@internal/constants';
 import chalk from 'chalk';
 import cors from 'cors';
 import {configDotenv} from 'dotenv';
@@ -6,12 +6,12 @@ import express, {type Request} from 'express';
 import minimist from 'minimist';
 import {join as joinPath} from 'path';
 
-import {fetchWithQuery, type FetchWithQueryParams} from './api';
+import {fetchWithQuery, type FetchWithQueryParams} from './fetch';
 
 configDotenv({path: joinPath(REPO_DIR, '.env')});
 
 const args = minimist(process.argv.slice(2));
-const port = parseInt(args.port, /* radix= */ 10) || SERVER_PORT;
+const port = parseInt(args.port, /* radix= */ 10) || API_PORT;
 
 const app = express();
 app.use(cors());
